backend/requirements.txt
# Core AI Frameworks
crewai==0.118.0
langchain>=0.3.21
langchain-community>=0.3.23
langchain-core>=0.3.45
langchain-openai>=0.1.0
huggingface-hub>=0.20.0

# Embeddings & Vector Search
sentence-transformers>=2.2.0
faiss-cpu==1.7.4
rank-bm25>=0.2.2
FlagEmbedding>=1.2.0

# Transformers
transformers>=4.42.0

# Document Processing
pypdf>=3.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0,<2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Visualization & Export
plotly>=5.14.0
openpyxl>=3.1.0